pyproject.toml
[tool.poetry]
name = "bookshare"
version = "0.1.0"
description = "Share books!"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
